Fixes index out of bounds for existing symbol #3129
